Handle missing TabletRelease event
After some significant investigating I found that tablet support in QML is fundamentally broken. I suspect it will stay that way until the Krita gets moved over to qtquick for it's canvas (which is planned).

That said the above is still a good fix to have for code correctness even if it doesn't get tablets working with Canonic.