Update what-is-qml.mdx fix minor typos (#4)
* Update what-is-qml.mdx fix minor typos

* Update what-is-qml.mdx more minor typos

I changed "it's" to "its". Added "." to etc.. (including ".." at end of sentence since I use Chicago's book of style).

Changed all "centre" to "center" because one of the comment lines in the code used the American spelling and it matches the code property spelling. (If prefer UK that's fine.)